refactor(components): [carousel]
* Extract component logic out of .vue file.
* Reorganize code for better readability.
