+++ packages/components/carousel/src/instance.ts
@@ -1,0 +1,3 @@
+import type Carousel from './carousel.vue'
+
+export type CarouselInstance = InstanceType<typeof Carousel>
